Encode event index names as integers (save space)
For each index defined for an event, we were encoding this as:

[uvarint:namelength][bytes(namelength):name][int64:previous][int64:next]

This means we were including the name of the index in event
metadata for each event in the index. This change switches the
event index metadata to the format:

[uvarint:indexKey][int64:previous][int64:next]

This saves a ton of space, and when looking up an index in the
block sst index, we now include the index key integer as well,
for easily finding the previous/next offset for an index when
scanning.