Index offsets are now stored sequentially.
Previous, all events were encoded to include previous
and next offsets for each index they were apart. This
lead to problems when attempting to allow compression,
as the value of those offsets were apart of the event,
and changed the length of the event when compressed.

Since the offsets could link to earlier or later in the
file, you couldn't really calculate an accurate offset,
as that my change the offset value in a previous event,
which changes the event's offset again. :)

Also, a doubly linking events was essentially storing
every offset twice, doubling the overhead.